+++ src/app/explorer/page.tsx
@@ -1,4 +1,4 @@
-import { AppSidebar } from "@/components/app-sidebar";
+import { AppSidebar } from "@/app/sidebar/app-sidebar";
 import AppBreadcrumb from "@/components/breadcrumb";
 import {} from "@/components/ui/breadcrumb";
 import { SidebarInset, SidebarProvider, SidebarTrigger } from "@/components/ui/sidebar";
