+++ src/app/sidebar/app-sidebar.tsx
@@ -10,9 +10,10 @@ import {
 } from "@/components/ui/sidebar";
 import { useAppDispatch, useAppSelector } from "@/hooks/store";
 import { selectHeaders, selectRootResources, setHeaders } from "@/state/store";
-import { Label } from "./ui/label";
-import { ResourceTypeList } from "@/app/sidebar/resource_type_list";
+import { Label } from "../../components/ui/label";
+import { ResourceTypeList } from "@/components/resource_types/resource_type_list";
 
+// The AppSidebar. This fetches the list of root resources from the schema and displays them.
 export function AppSidebar({ ...props }: React.ComponentProps<typeof Sidebar>) {
   const resources = useAppSelector(selectRootResources);
